wp-members-pluggable.php
These are rewrites of the functions which govern output of most of the forms for the WP-Members
plugin for use with Twitter Bootstrap.  Note, this code uses pluggable functions, and should be placed
in either your theme's functions.php file or the plugins wp-members-pluggable.php file.  There is
a chance that versions of the plugin later than 2.9.9.1 may not work with this method.